stock is in progress

+++ src/main/java/com/pems/backend/entity/Product.java
@@ -3,7 +3,7 @@ package com.pems.backend.entity;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-import jakarta.persistence.Column;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -26,7 +26,7 @@ public class Product {
 
 	private String productCode;
 
-	private Integer quantity;
+	private Long quantity;
 
 	private BigDecimal purchasePrice;
